HorizontalBarSection.tsx convert for OverlayTrigger and Popover - and convert test to RTL
