+++ packages/components/src/theme/charts.scss
@@ -42,6 +42,7 @@
     display: inline-block;
     position: relative;
     height: 20px;
+    width: 100%;
     outline: 1px solid $text-color-gray;
     vertical-align: bottom;
 }
